+++ dimas/src/com/queryable.rs
@@ -32,15 +32,11 @@ impl Request {
 	///
 	pub fn reply<T>(&self, value: T)
 	where
-		T: bincode::Encode,
+		T: bitcode::Encode,
 	{
 		let key = self.query.selector().key_expr.to_string();
-		let sample = Sample::try_from(
-			key,
-			bincode::encode_to_vec(&value, bincode::config::standard())
-				.expect("should never happen"),
-		)
-		.expect("should never happen");
+		let encoded: Vec<u8> = bitcode::encode(&value).expect("should never happen");
+		let sample = Sample::try_from(key, encoded).expect("should never happen");
 
 		self.query
 			.reply(Ok(sample))